Fully working API Console on Get Started page, updated build scripts to compile the additional React app
