--- _includes/ava_console.html
@@ -1,184 +1,0 @@
-<link rel="stylesheet" type="text/css" href="/css/ava_console.css">
-<div class="loading">Loading&#8230;</div>
-
-<h3>Validate Address</h3>
-<table>
-    <tr>
-        <td><strong>Description</strong></td>
-        <td>Validate an address</td>
-    </tr>
-    <tr>
-        <td><strong>Endpoint</strong></td>
-        <td>https://developer.avalara.com/1.0/address/validate</td>
-    </tr>
-    <tr>
-        <td><strong>HTTP Method</strong></td>
-        <td>GET</td>
-    </tr>
-</table>
-<form id="validateAddress">
-    <h4>Request</h4>
-    <table>
-        <tr>
-            <td><label for="Line1">Line1*</label></td>
-            <td><input id="Line1" name="Line1" data-required="true" placeholder="1100 2nd Ave" /></td>
-        </tr>
-        <tr>
-            <td><label for="Line2">Line2</label></td>
-            <td><input id="Line2" name="Line2" placeholder="Suite 300" /></td>
-        </tr>
-        <tr>
-            <td><label for="Line3">Line3</label></td>
-            <td><input id="Line3" name="Line3"></td>
-        </tr>
-        <tr>
-            <td><label for="City">City*</label></td>
-            <td><input id="City" name="City" data-required="true" placeholder="Seattle"></td>
-        </tr>
-        <tr>
-            <td><label for="Region">Region</label></td>
-            <td><input id="Region" name="Region"></td>
-        </tr>
-        <tr>
-            <td><label for="Country">Country*</label></td>
-            <td><input id="Country" name="Country" data-required="true" placeholder="United States"></td>
-        </tr>
-        <tr>
-            <td><label for="PostalCode">PostalCode*</label></td>
-            <td><input id="PostalCode" name="PostalCode" data-required placeholder="98134"></td>
-        </tr>
-    </table>
-    <br />
-    <button class="btn btn-success" type="submit">Submit</button>
-    <button class="btn btn-default" id="fillSampleAddressData">Fill Sample Data</button>
-    <button class="btn btn-default" type="reset">Reset</button>
-    </form>
-    <br />
-    <div id="validateAddressResponse">
-        <table>
-            <tr>
-                <td><strong>HTTP Response Code</strong></td>
-                <td id='validate-address-status' class='code-snippet-code-text'></td>
-            </tr>
-            <tr>
-                <td style="vertical-align:top"><strong>HTTP Response Body</strong></td>
-                <td class='code-snippet-code-text'>
-                    <textarea style="background: dimgrey" class='responseBody' rows="15" cols="50"></textarea>
-                </td>
-            </tr>
-        </table>
-    </div>
-    <h3>Get Tax</h3>
-    <table>
-        <tr>
-            <td><strong>Description</strong></td>
-            <td>Calculates taxes on a document such as a sales order, sales invoice or purchase order</td>
-        </tr>
-        <tr>
-            <td><strong>Endpoint</strong></td>
-            <td>https://developer.avalara.com/1.0/tax/get</td>
-        </tr>
-        <tr>
-            <td><strong>HTTP Method</strong></td>
-            <td>POST</td>
-        </tr>
-        </table>
-        <form id="getTax">
-            <h4>Request</h4>
-            <table>
-        <tr id="gtRequest">
-            <td colspan="2">
-                <table>
-                    <tr>
-                        <td><label for="DocDate">DocDate</label></td>
-                        <td><input id="DocDate" name="DocDate" data-required="true" placeholder="2016-06-01"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="CustomerCode">CustomerCode</label></td>
-                        <td><input id="CustomerCode" name="CustomerCode" data-required="true" placeholder="0000"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="CompanyCode">CompanyCode</label></td>
-                        <td><input id="CompanyCode" name="CompanyCode" data-required="true" placeholder="APITrialCompany"></td>
-                    </tr>
-                </table>
-            </td>
-        </tr>
-        <tr>
-            <td colspan="2"><strong>Addresses</strong></td>
-        </tr>
-        <tr id="gtRequestAddresses">
-            <td colspan="2">
-                <table>
-                    <tr>
-                        <td><label for="AddressCode">AddressCode</label></td>
-                        <td><input id="AddressCode" name="AddressCode" data-required="true" placeholder="1"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="gtLine1">Line1</label></td>
-                        <td><input id="gtLine1" name="Line1" data-required="true" placeholder="435 Ericksen Avenue Northeast"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="gtLine2">Line2</label></td>
-                        <td><input id="gtLine2" name="Line2" data-required="true" placeholder="#250"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="gtPostalCode">PostalCode</label></td>
-                        <td><input id="gtPostalCode" name="PostalCode" data-required="true" placeholder="98110"></td>
-                    </tr>
-                </table>
-            </td>
-        </tr>
-        <tr>
-            <td colspan="2"><strong>Lines</strong></td>
-        </tr>
-        <tr id="gtRequestLines">
-            <td colspan="2">
-                <table>
-                    <tr>
-                        <td><label for="LineNo">LineNo</label></td>
-                        <td><input id="LineNo" name="LineNo" data-required="true" placeholder="1"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="DestinationCode">DestinationCode</label></td>
-                        <td><input id="DestinationCode" name="DestinationCode" data-required="true" placeholder="1"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="OriginCode">OriginCode</label></td>
-                        <td><input id="OriginCode" name="OriginCode" data-required="true" placeholder="1"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="Qty">Qty</label></td>
-                        <td><input id="Qty" name="Qty" data-required="true" placeholder="1"></td>
-                    </tr>
-                    <tr>
-                        <td><label for="Amount">Amount</label></td>
-                        <td><input id="Amount" name="Amount" data-required="true" placeholder="10"></td>
-                    </tr>
-                </table>
-            </td>
-        </tr>
-    </table>
-    <br />
-    <button class="btn btn-success" type="submit">Submit</button>
-    <button class="btn btn-default" id="fillGetTaxSamplePost">Fill Sample Data</button>
-    <button class="btn btn-default" type="reset">Reset</button>
-    </form>
-    <br/>
-    <div id="getTaxResponse">
-        <table>
-            <tr>
-                <td><strong>HTTP Response Code</strong></td>
-                <td id='get-tax-status' class='code-snippet-code-text'></td>
-            </tr>
-            <tr>
-                <td style="vertical-align:top"><strong>HTTP Response Body</strong></td>
-                <td class='code-snippet-code-text'>
-                    <textarea style="background: dimgrey" class='responseBody' rows="15" cols="50"></textarea>
-                </td>
-            </tr>
-        </table>
-    </div>
-
-<script src="https://code.jquery.com/jquery-3.0.0.min.js" integrity="sha256-JmvOoLtYsmqlsWxa7mDSLMwa6dZ9rrIdtrrVYRnDRH0=" crossorigin="anonymous"></script>
-<script src="/js/ava_console.js"></script>
