+++ _includes/get-started-console.html
@@ -1,0 +1,3 @@
+<div id='react-root'>
+</div>
+<script src='/dynamic/public/javascript/build/get-started-bundle.js'></script>
