+++ _sass/_api-console.scss
@@ -17,6 +17,10 @@
     padding: 5px;
 }
 
+.m-l-1 {
+    margin-left: 1rem;
+}
+
 .code-snippet {
     border: 1px solid #71e4ff;
     padding: 10px;
