+++ css/console.css
@@ -1,7 +1,3 @@
-.m-l-1 {
-    margin-left: 1rem;
-}
-
 .responseBody {
     margin-top: 1rem;
 }
