+++ dynamic/react/shared/reducers/queryStringReducer.js
